requirements.txt
pydantic
motor
fastapi
uvicorn
python-dotenv